requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
python-dateutil==2.8.2
google-generativeai==0.3.0
protobuf==4.25.3
tqdm==4.66.4
google-ai-generativelanguage==0.4.0
google-api-core>=2.11.0
google-auth>=2.17.0
typing-extensions>=4.9.0
